backend/requirements/dev.txt
-r base.txt

black==26.3.1
cfgv==3.5.0
debugpy==1.8.20
distlib==0.4.0
django-debug-toolbar==6.3.0
django-extensions==4.1
docker==7.1.0
filelock==3.25.2
flake8==7.3.0
identify==2.6.18
iniconfig==2.3.0
isort==8.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.10.0
pathspec==1.0.4
platformdirs==4.9.6
pluggy==1.6.0
pre_commit==4.5.1
psutil==7.2.2
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.20.0
pytest==9.0.3
pytest-celery==1.3.0
pytest-django==4.12.0
pytest_docker_tools==3.1.9
python-crontab==3.3.0
python-discovery==1.2.2
pytokens==0.4.1
tenacity==9.1.4
virtualenv==21.2.1
wcwidth==0.6.0